Migrate `DefaultPerformancePointsCounter` and rename it
